# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/fx_gain_losses:
    get:
      tags:
        - FxGainLosses
      summary: List FxGainLosses
      description: Get a paginated list of FxGainLosses with optional filtering and sorting
      operationId: listFxGainLosses
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FxGainLossList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - FxGainLosses
      summary: Create FxGainLoss
      description: Create a new FxGainLoss
      operationId: createFxGainLoss
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFxGainLossRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FxGainLoss'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    FxGainLoss:
      type: object
      required:
        - id
        - company_id
        - matched_source_id
        - currency
        - original_rate
        - realised_rate
        - base_amount_delta
        - direction
        - fx_account_id
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        bank_clearance_id:
          nullable: true
          type: string
          format: uuid
        matched_source_id:
          type: string
          format: uuid
        currency:
          type: string
          maxLength: 3
        original_rate:
          type: string
        realised_rate:
          type: string
        base_amount_delta:
          type: string
        direction:
          type: string
        fx_account_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateFxGainLossRequest:
      type: object
      required:
        - company_id
        - matched_source_id
        - currency
        - original_rate
        - realised_rate
        - base_amount_delta
        - direction
        - fx_account_id
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        bank_clearance_id:
          nullable: true
          type: string
          format: uuid
        matched_source_id:
          type: string
          format: uuid
        currency:
          type: string
          maxLength: 3
        original_rate:
          type: string
        realised_rate:
          type: string
        base_amount_delta:
          type: string
        direction:
          type: string
        fx_account_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    FxGainLossList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/FxGainLoss'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []