# src/presentation/http/fx_gain_loss_handler.rs
//! FxGainLoss REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;
use rust_decimal::Decimal;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{FxGainLossService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateFxGainLossDto, UpdateFxGainLossDto, PatchFxGainLossDto, FxGainLossResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum FxGainLossError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for FxGainLossError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for FxGainLossError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "FXGAINLOSS_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "FXGAINLOSS_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "FXGAINLOSS_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "FXGAINLOSS_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for FxGainLoss.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /fx_gain_losses | List with pagination |
/// | POST | /fx_gain_losses | Create new |
/// | GET | /fx_gain_losses/:id | Get by ID |
/// | PUT | /fx_gain_losses/:id | Full update |
/// | PATCH | /fx_gain_losses/:id | Partial update |
/// | DELETE | /fx_gain_losses/:id | Soft delete |
/// | POST | /fx_gain_losses/bulk | Bulk create |
/// | POST | /fx_gain_losses/upsert | Upsert |
/// | GET | /fx_gain_losses/trash | List deleted |
/// | POST | /fx_gain_losses/:id/restore | Restore |
/// | DELETE | /fx_gain_losses/empty | Empty trash |
/// | GET | /fx_gain_losses/:id/deleted | Get deleted by ID |
/// | DELETE | /fx_gain_losses/trash/:id | Permanent delete from trash |
/// | GET | /fx_gain_losses/count | Count active entities |
/// | GET | /fx_gain_losses/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(FxGainLossService::with_repository(repository));
/// let router = create_fx_gain_loss_routes(service);
/// ```
pub fn create_fx_gain_loss_routes(service: Arc<FxGainLossService>) -> Router {
    BackboneCrudHandler::<FxGainLossService, FxGainLoss, CreateFxGainLossDto, UpdateFxGainLossDto, FxGainLossResponseDto>::routes(
        service,
        "/fx_gain_losses",
    )
}

/// Create Axum router with only the read (GET) endpoints for FxGainLoss.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_fx_gain_loss_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_fx_gain_loss_read_routes(service: Arc<FxGainLossService>) -> Router {
    BackboneCrudHandler::<FxGainLossService, FxGainLoss, CreateFxGainLossDto, UpdateFxGainLossDto, FxGainLossResponseDto>::read_routes(
        service,
        "/fx_gain_losses",
    )
}

/// Create Axum router with only the write (mutation) endpoints for FxGainLoss.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
pub fn create_fx_gain_loss_write_routes(service: Arc<FxGainLossService>) -> Router {
    BackboneCrudHandler::<FxGainLossService, FxGainLoss, CreateFxGainLossDto, UpdateFxGainLossDto, FxGainLossResponseDto>::write_routes(
        service,
        "/fx_gain_losses",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_fx_gain_loss_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<FxGainLossService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_fx_gain_loss_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}